requirements.txt
# Intelligent Job Scraper Dependencies
# Only the essential packages needed for the modular scraper

# AI Integration
google-generativeai>=0.3.0

# Web Scraping and Browser Automation
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Environment Configuration
python-dotenv>=1.0.0
